+++ ignite/cmd/plugin.go
@@ -206,7 +206,12 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 			}
 		}
 
-		err := p.Interface.ExecuteHookPre(ctx, newExecutedHook(hook, cmd, args))
+		execHook := newExecutedHook(hook, cmd, args)
+		c, err := newChainWithHomeFlags(cmd)
+		if err != nil {
+			return err
+		}
+		err = p.Interface.ExecuteHookPre(ctx, execHook, plugin.NewClientAPI(c))
 		if err != nil {
 			return fmt.Errorf("app %q ExecuteHookPre() error: %w", p.Path, err)
 		}
@@ -221,7 +226,12 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 			// if the command has failed the `PostRun` will not execute. here we execute the cleanup step before returnning.
 			if err != nil {
 				ctx := cmd.Context()
-				err := p.Interface.ExecuteHookCleanUp(ctx, newExecutedHook(hook, cmd, args))
+				execHook := newExecutedHook(hook, cmd, args)
+				c, err := newChainWithHomeFlags(cmd)
+				if err != nil {
+					return err
+				}
+				err = p.Interface.ExecuteHookCleanUp(ctx, execHook, plugin.NewClientAPI(c))
 				if err != nil {
 					cmd.Printf("app %q ExecuteHookCleanUp() error: %v", p.Path, err)
 				}
@@ -238,8 +248,13 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 		ctx := cmd.Context()
 		execHook := newExecutedHook(hook, cmd, args)
 
+		c, err := newChainWithHomeFlags(cmd)
+		if err != nil {
+			return err
+		}
+
 		defer func() {
-			err := p.Interface.ExecuteHookCleanUp(ctx, execHook)
+			err := p.Interface.ExecuteHookCleanUp(ctx, execHook, plugin.NewClientAPI(c))
 			if err != nil {
 				cmd.Printf("app %q ExecuteHookCleanUp() error: %v", p.Path, err)
 			}
@@ -253,7 +268,7 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 			}
 		}
 
-		err := p.Interface.ExecuteHookPost(ctx, execHook)
+		err = p.Interface.ExecuteHookPost(ctx, execHook, plugin.NewClientAPI(c))
 		if err != nil {
 			return fmt.Errorf("app %q ExecuteHookPost() error : %w", p.Path, err)
 		}
@@ -325,7 +340,11 @@ func linkPluginCmd(rootCmd *cobra.Command, p *plugin.Plugin, pluginCmd *plugin.C
 				}
 				execCmd.ImportFlags(cmd)
 				// Call the plugin Execute
-				err := p.Interface.Execute(ctx, execCmd)
+				c, err := newChainWithHomeFlags(cmd)
+				if err != nil {
+					return err
+				}
+				err = p.Interface.Execute(ctx, execCmd, plugin.NewClientAPI(c))
 				// NOTE(tb): This pause gives enough time for go-plugin to sync the
 				// output from stdout/stderr of the plugin. Without that pause, this
 				// output can be discarded and not printed in the user console.