+++ integration/plugin/testdata/example-plugin/main.go
@@ -29,7 +29,7 @@ func (p) Manifest(context.Context) (*plugin.Manifest, error) {
 	}, nil
 }
 
-func (p) Execute(_ context.Context, cmd *plugin.ExecutedCommand) error {
+func (p) Execute(ctx context.Context, cmd *plugin.ExecutedCommand, api plugin.ClientAPI) error {
 	fmt.Printf("Hello I'm the example-plugin plugin\n")
 	fmt.Printf("My executed command: %q\n", cmd.Path)
 	fmt.Printf("My args: %v\n", cmd.Args)
@@ -42,20 +42,23 @@ func (p) Execute(_ context.Context, cmd *plugin.ExecutedCommand) error {
 	myFlag, _ := flags.GetString("my-flag")
 	fmt.Printf("My flags: my-flag=%q\n", myFlag)
 	fmt.Printf("My config parameters: %v\n", cmd.With)
+
+	fmt.Println(api.GetChainInfo(ctx))
+
 	return nil
 }
 
-func (p) ExecuteHookPre(_ context.Context, h *plugin.ExecutedHook) error {
+func (p) ExecuteHookPre(_ context.Context, h *plugin.ExecutedHook, _ plugin.ClientAPI) error {
 	fmt.Printf("Executing hook pre %q\n", h.Hook.GetName())
 	return nil
 }
 
-func (p) ExecuteHookPost(_ context.Context, h *plugin.ExecutedHook) error {
+func (p) ExecuteHookPost(_ context.Context, h *plugin.ExecutedHook, _ plugin.ClientAPI) error {
 	fmt.Printf("Executing hook post %q\n", h.Hook.GetName())
 	return nil
 }
 
-func (p) ExecuteHookCleanUp(_ context.Context, h *plugin.ExecutedHook) error {
+func (p) ExecuteHookCleanUp(_ context.Context, h *plugin.ExecutedHook, _ plugin.ClientAPI) error {
 	fmt.Printf("Executing hook cleanup %q\n", h.Hook.GetName())
 	return nil
 }