+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Features
+
+- [#3544](https://github.com/ignite/cli/pull/3544) Add bidirectional communication to plugin system
+- [#3561](https://github.com/ignite/cli/pull/3561) Add GetChainInfo method to plugin system API
+
 ### Changes
 
 - [#3529](https://github.com/ignite/cli/pull/3529) Refactor plugin system to use gRPC